move store types to ts files
