+++ src/components/Graph/actions.ts
@@ -6,7 +6,7 @@ import {
   QueryTarget,
   GraphValueAxes,
   GraphValuesListener,
-} from '@/store/types';
+} from '@/store/history';
 import parseDuration from 'parse-duration';
 import { nanoToMilli } from '@/helpers/functional';
 import forEach from 'lodash/forEach';
