+++ src/plugins/history/Graph/index.ts
@@ -1,5 +1,5 @@
 import { ref } from '@/helpers/component-ref';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import { typeName } from './getters';
 import form from './GraphForm.vue';
 import widget from './GraphWidget.vue';
