+++ src/plugins/history/Graph/types.d.ts
@@ -1,5 +1,5 @@
 import { GraphConfig } from '@/components/Graph/types';
-import { DashboardItem } from '@/store/types';
+import { DashboardItem } from '@/store/dashboards';
 
 export interface HistoryItem extends DashboardItem {
   config: GraphConfig;
