+++ src/plugins/history/Metrics/MetricsWidget.vue
@@ -6,7 +6,7 @@ import historyStore from '@/store/history';
 import get from 'lodash/get';
 import parseDuration from 'parse-duration';
 import { durationString } from '@/helpers/functional';
-import { QueryTarget, Listener, DisplayNames, QueryParams } from '@/store/types';
+import { QueryTarget, Listener, DisplayNames, QueryParams } from '@/store/history';
 import { MetricsResult, MetricsConfig } from './types';
 import { DEFAULT_FRESH_DURATION, DEFAULT_DECIMALS } from './getters';
 import { addListener } from './actions';
