+++ src/plugins/history/Metrics/index.ts
@@ -1,5 +1,5 @@
 import { ref } from '@/helpers/component-ref';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import widget from './MetricsWidget.vue';
 import wizard from './MetricsWizard.vue';
 import form from './MetricsForm.vue';
