+++ src/plugins/spark/arrangements/BrewPi/BrewPiSettingsTask.vue
@@ -17,7 +17,7 @@ import {
   typeName as pidType,
   defaultData as pidData,
 } from '@/plugins/spark/features/Pid/getters';
-import { Dashboard } from '@/store/types';
+import { Dashboard } from '@/store/dashboards';
 
 @Component
 export default class BrewPiSettingsTask extends WizardTaskBase {
