+++ src/plugins/spark/arrangements/BrewPi/types.d.ts
@@ -1,5 +1,5 @@
 import { Block } from '@/plugins/spark/types';
-import { DashboardItem } from '@/store/types';
+import { DashboardItem } from '@/store/dashboards';
 
 export interface BrewPiConfigNames {
   fridgeSensor: string;
