+++ src/plugins/spark/arrangements/index.ts
@@ -1,4 +1,4 @@
-import { Arrangement } from '@/store/types';
+import { Arrangement } from '@/store/features';
 import BrewPi from './BrewPi';
 
 const arrangements: { [id: string]: Arrangement } = {
