+++ src/plugins/spark/features/ActuatorDS2413/index.ts
@@ -1,6 +1,6 @@
 import { ref } from '@/helpers/component-ref';
 import GenericBlock from '@/plugins/spark/components/GenericBlock';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import form from './ActuatorDS2413Form.vue';
 import widget from './ActuatorDS2413Widget.vue';
 import { typeName } from './getters';
