+++ src/plugins/spark/features/ActuatorPwm/index.ts
@@ -1,6 +1,6 @@
 import { ref } from '@/helpers/component-ref';
 import GenericBlock from '@/plugins/spark/components/GenericBlock';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import form from './ActuatorPwmForm.vue';
 import widget from './ActuatorPwmWidget.vue';
 import { typeName } from './getters';
