+++ src/plugins/spark/features/DS2413/index.ts
@@ -1,6 +1,6 @@
 import { ref } from '@/helpers/component-ref';
 import GenericBlock from '@/plugins/spark/components/GenericBlock';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import form from './DS2413Form.vue';
 import widget from './DS2413Widget.vue';
 import { typeName } from './getters';
