+++ src/plugins/spark/features/SessionView/index.ts
@@ -1,5 +1,5 @@
 import { ref } from '@/helpers/component-ref';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import widget from './SessionViewWidget.vue';
 import form from './SessionViewForm.vue';
 import wizard from './SessionViewWizard.vue';
