+++ src/plugins/spark/features/SetpointProfile/index.ts
@@ -1,6 +1,6 @@
 import { ref } from '@/helpers/component-ref';
 import GenericBlock from '@/plugins/spark/components/GenericBlock';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import { typeName } from './getters';
 import form from './SetpointProfileForm.vue';
 import widget from './SetpointProfileWidget.vue';
