+++ src/plugins/spark/features/TempSensorMock/index.ts
@@ -1,6 +1,6 @@
 import { ref } from '@/helpers/component-ref';
 import GenericBlock from '@/plugins/spark/components/GenericBlock';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import { typeName } from './getters';
 import form from './TempSensorMockForm.vue';
 import widget from './TempSensorMockWidget.vue';
