+++ src/plugins/spark/features/TempSensorOneWire/index.ts
@@ -1,6 +1,6 @@
 import { ref } from '@/helpers/component-ref';
 import GenericBlock from '@/plugins/spark/components/GenericBlock';
-import { Feature } from '@/store/types';
+import { Feature } from '@/store/features';
 import { typeName } from './getters';
 import form from './TempSensorOneWireForm.vue';
 import widget from './TempSensorOneWireWidget.vue';
