+++ src/plugins/spark/index.ts
@@ -14,7 +14,7 @@ import {
   truncate,
 } from '@/helpers/functional';
 import { Link, Unit } from '@/helpers/units';
-import { Service } from '@/store/types';
+import { Service } from '@/store/services';
 import features from './features';
 import arrangements from './arrangements';
