+++ src/plugins/spark/provider/SparkPage.vue
@@ -3,11 +3,11 @@ import Vue from 'vue';
 import Component from 'vue-class-component';
 import { uid, Dialog } from 'quasar';
 import dashboardStore from '@/store/dashboards';
-import featureStore from '@/store/features';
+import featureStore, { FeatureRole } from '@/store/features';
 import serviceStore from '@/store/services';
 import sparkStore from '@/plugins/spark/store';
 import { Block, SystemStatus, Spark } from '@/plugins/spark/types';
-import { Dashboard, DashboardItem, FeatureRole } from '@/store/types';
+import { Dashboard, DashboardItem } from '@/store/dashboards';
 import { isReady, isSystemBlock, widgetSize } from './getters';
 import { Watch } from 'vue-property-decorator';
 import { setInterval, clearTimeout } from 'timers';
