+++ src/store/dashboards/index.ts
@@ -1,7 +1,6 @@
 import Vue from 'vue';
 import store from '@/store';
 import { Module, VuexModule, Mutation, Action, getModule } from 'vuex-module-decorators';
-import { Dashboard, DashboardItem } from '@/store/types';
 import {
   createDashboard as createDashboardInApi,
   createDashboardItem as createDashboardItemInApi,
@@ -14,6 +13,27 @@ import {
   setupDashboards as setupDashboardsInApi,
   setupDashboardItems as setupDashboardItemsInApi,
 } from './api';
+
+export interface Dashboard {
+  id: string;
+  title: string;
+  order: number;
+  primary?: boolean;
+  _rev?: string;
+}
+
+export interface DashboardItem {
+  id: string;
+  title: string;
+  cols: number;
+  rows: number;
+  order: number;
+  dashboard: string;
+  feature: string;
+  config: any;
+  pinnedPosition?: XYPosition | null;
+  _rev?: string;
+}
 
 @Module({ store, namespaced: true, dynamic: true, name: 'dashboards' })
 export class DashboardModule extends VuexModule {
