+++ src/store/history/index.ts
@@ -1,13 +1,60 @@
 import Vue from 'vue';
 import store from '@/store';
 import { Module, VuexModule, Mutation, Action, getModule } from 'vuex-module-decorators';
-import { Listener, QueryResult, QueryParams, QueryTarget } from '@/store/types';
 import {
   fetchKnownKeys as fetchKnownKeysInApi,
   validateService as validateServiceInApi,
   subscribeMetrics,
   subscribeValues,
 } from './api';
+
+export interface QueryParams {
+  database?: string;
+  start?: string | number;
+  duration?: string;
+  end?: string | number;
+  limit?: number;
+  orderBy?: string;
+  approxPoints?: number;
+}
+
+export interface QueryTarget {
+  measurement: string;
+  fields: string[];
+}
+
+export interface DisplayNames {
+  [key: string]: string;
+}
+
+export interface GraphValueAxes {
+  [key: string]: 'y' | 'y2';
+}
+
+export type Slice = number[];
+
+export interface QueryResult {
+  name: string;
+  columns: string[];
+  values: Slice[];
+  database: string;
+  policy: string;
+}
+
+export interface Listener {
+  id: string;
+  transformer: (listener: any, result: any) => Listener;
+  params: QueryParams;
+  target: QueryTarget;
+  renames: DisplayNames;
+  source?: EventSource;
+  values?: any;
+}
+
+export interface GraphValuesListener extends Listener {
+  axes: GraphValueAxes;
+  usedPolicy?: string;
+}
 
 
 @Module({ store, namespaced: true, dynamic: true, name: 'history' })
